Add JsonFormatter VS library project
