+++ msvc-full-features/JsonFormatter-vcpkg-static.vcxproj
@@ -58,7 +58,7 @@
   </ImportGroup>
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup>
-    <TargetName>json_formatter</TargetName>
+    <TargetName>$(ProjectName)-$(Configuration)-$(Platform)</TargetName>
     <TargetExt>.exe</TargetExt>
     <OutDir>$(SolutionDir)..\tools\format\</OutDir>
     <IntDir>$(SolutionDir)$(ProjectName)\$(Configuration)\$(Platform)\</IntDir>
@@ -131,7 +131,7 @@
     <ClInclude Include="stdafx.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="..\tools\format\*.cpp" />
+    <ClCompile Include="..\tools\format\format_main.cpp" />
     <ClCompile Include="..\src\messages.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
@@ -140,6 +140,9 @@
   <ItemGroup>
     <ProjectReference Include="Cataclysm-lib-vcpkg-static.vcxproj">
       <Project>{0009bb11-11ad-4c14-a5fc-d882a942c00b}</Project>
+    </ProjectReference>
+    <ProjectReference Include="JsonFormatter-lib-vcpkg-static.vcxproj">
+      <Project>{534a4e38-96a1-40e4-bda7-8d17607f0270}</Project>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
