+++ tools/format/CMakeLists.txt
@@ -3,6 +3,7 @@ include(ExternalProject)
 add_executable(
     json_formatter
     format.cpp
+    format_main.cpp
     ${CMAKE_SOURCE_DIR}/src/json.cpp
     )
